Bug fix + Refactoring
Resolved issue where you were unable to promote pieces by clicking to move to the back rank squares, for built in react-chessboard promotion ui you have to drag and drop the piece to that square. Also starting to refactor ChessGame.tsx to reduce clutter will start to create more util files in the future to reduce length of this main file.