State Centralization in play/page.tsx
Moved fenHistory / setFenHistory and activePlayer / setActivePlayer to the parent component to have that be the one true source of truth and not to create conflicting states.

Ability to take back moves now works throughout the app since I play an adding primarily AI opponents I will not be adding conformation for the takeback moves for now since the AI should always allow move takebacks.