Updated elo 1320-3190
Stockfish can only accurately immeditate 1320-3190 elo players so adjusted difficulty ratings.

Added last move highlighting to let user know which piece was last to move and where it moved from.

Another useeffect to check to see if the length of the fenhistory got shorter by 1 ( a move was taken back ) and then clears the highlighting to not be misleading and confusing.