Refactor useEffects, Custom Hooks
useChessGame custom hook added to parent page.tsx component for the main /play route.

Refactored useEffects in ChessGame.tsx to combine effects that relied on same dependancies to improve readability and performance.

CalculateMaterialDifference util file for the ChessGame.tsx to pass down to playinfo component when displaying the material difference in the games.