Building on puzzles functionality
Currently the puzzles work if you are making moves for both sides of the board.

Still need to improve the function for the computer making the every other move in each puzzle always starting with the first move.

Also still need to add proper functionality for clicking to move pieces in my handleSquareClickCallback function in my PuzzleGame.tsx.

Lastly will need to add more ui improvements for the puzzle functionality and better filtering for puzzles based on theme selected in the PuzzleInitializer currently just default values for development.