UI Changes + Improved confirmation dialogs.
Improved responsiveness of various UI components based on screen size to ensure everything always looks good.

Added custom confirmation dialog box for skipping puzzles and when you run out of puzzles and are about to be redirected back to the theme selection screen.

Added hinthighlighting styles for when a user requests a hint in the puzzle component.

--removed react-toastify from my project, no need for it anymore currently.